# medtriage/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MedTriageAI API
  version: 1.0.0
  description: |
    Cloud-based AI-assisted chest X-ray triage API.

    **Non-diagnostic, decision-support only.** All responses are intended to
    aid a qualified clinician's review and must not be used as a sole basis
    for diagnosis or treatment.

    Pipeline overview:
      1. Client uploads a chest X-ray to `/upload`.
      2. The image is stored in Cloud Storage and a case document is created
         in Firestore with `status = pending`.
      3. An asynchronous worker chain runs Cloud Vision + CheXNet, then
         Gemini 2.5 Flash, updating the case status through
         `vision_done → analyzing → complete`.
      4. A clinician reviews the case at `/cases/{caseId}/review` and signs off.
      5. A PDF report is rendered and emailed to the patient.
  contact:
    name: MedTriageAI Team
    email: contact@medtriage.ai
  license:
    name: Proprietary
servers:
  - url: https://api.medtriage.ai/v1
    description: Production (via API Gateway)
  - url: https://medtriage.ew.r.appspot.com/api
    description: App Engine direct (MVP / staging)
...
security:
  - bearerAuth: []
paths:
...
  /cases:
    get:
      tags: [Cases]
      summary: List cases for the authenticated user / tenant
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/CaseStatus'
          description: Optional filter on case status
        - in: query
          name: severity
          schema:
            $ref: '#/components/schemas/Severity'
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: pageToken
          schema:
            type: string
      responses:
        '200':
          description: Page of cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseSummary'
                  nextPageToken:
                    type: string
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cases/{caseId}:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'
    get:
      tags: [Cases]
      summary: Retrieve a case
      responses:
        '200':
          description: Case document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backend session JWT issued by `/auth/session`
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    mtlsClientCert:
      type: mutualTLS
      description: PACS-side mTLS for DICOM ingestion
  parameters:
    CaseIdParam:
      in: path
      name: caseId
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    CaseStatus:
      type: string
      enum:
        - pending
        - vision_processing
        - vision_done
        - analyzing
        - complete
        - awaiting_radiologist
        - signed_off
        - error
    Severity:
      type: string
      enum: [normal, mild, moderate, severe, unknown]
...
    VisionLabel:
      type: object
      properties:
        description:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
    Finding:
      type: object
      properties:
        name:
          type: string
          description: CheXNet class name
        score:
          type: number
          minimum: 0
          maximum: 1
    Condition:
      type: object
      properties:
        name:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        location:
          type: string
          description: e.g. "right lower lobe", "bilateral"
        chexnetScore:
          type: number
          minimum: 0
          maximum: 1
    CaseSummary:
      type: object
      required: [caseId, status, timestamp]
      properties:
        caseId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/CaseStatus'
        classification:
          type: string
          nullable: true
        severity:
          $ref: '#/components/schemas/Severity'
        imageUrl:
          type: string
          format: uri
    Case:
      allOf:
        - $ref: '#/components/schemas/CaseSummary'
        - type: object
          properties:
            tenantId:
              type: string
            heatmap:
              type: string
              description: Base64-encoded PNG heatmap from Grad-CAM
              nullable: true
            visionLabels:
              type: array
              items:
                $ref: '#/components/schemas/VisionLabel'
            chexnetScores:
              type: object
              additionalProperties:
                type: number
                minimum: 0
                maximum: 1
            chexnetTopFindings:
              type: array
              items:
                $ref: '#/components/schemas/Finding'
            chexnetGradcamClass:
              type: string
              nullable: true
            conditions:
              type: array
              items:
                $ref: '#/components/schemas/Condition'
            findings:
              type: string
              description: Detailed radiological findings paragraph
            explanation:
              type: string
              description: Plain-language explanation for a non-radiologist
            error:
              type: string
              nullable: true
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true